+++ tests/test_httpx_async.py
@@ -464,11 +464,11 @@ async def test_requests_retrieval(httpx_mock: HTTPXMock):
     )
 
     async with httpx.AsyncClient() as client:
-        await client.post("http://test_url", data=b"sent content 2")
+        await client.post("http://test_url", content=b"sent content 2")
         await client.get("http://test_url", headers={"X-TEST": "test header 1"})
-        await client.put("http://test_url", data=b"sent content 3")
+        await client.put("http://test_url", content=b"sent content 3")
         await client.head("http://test_url")
-        await client.patch("http://test_url", data=b"sent content 5")
+        await client.patch("http://test_url", content=b"sent content 5")
         await client.delete("http://test_url", headers={"X-Test": "test header 4"})
 
     assert (
@@ -609,12 +609,15 @@ async def test_requests_json_body(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.post("http://test_url")
         assert response.json() == {"key 1": "value 1", "key 2": "value 2"}
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.get("http://test_url")
         assert response.json() == ["list content 1", "list content 2"]
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.put("http://test_url")
         assert response.json() == "string value"
+        assert response.headers["content-type"] == "application/json"
 
 
 @pytest.mark.asyncio
@@ -642,6 +645,7 @@ async def test_callback_returning_response(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.get("http://test_url")
         assert response.json() == {"url": "http://test_url"}
+        assert response.headers["content-type"] == "application/json"
 
 
 @pytest.mark.asyncio
@@ -654,9 +658,11 @@ async def test_callback_executed_twice(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.get("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.post("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
 
 @pytest.mark.asyncio
@@ -669,9 +675,11 @@ async def test_callback_matching_method(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.get("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.get("http://test_url2")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
 
 def test_request_retrieval_with_more_than_one(testdir):
@@ -743,7 +751,7 @@ async def test_content_matching(httpx_mock: HTTPXMock):
     httpx_mock.add_response(match_content=b"This is the body")
 
     async with httpx.AsyncClient() as client:
-        response = await client.post("http://test_url", data=b"This is the body")
+        response = await client.post("http://test_url", content=b"This is the body")
         assert response.read() == b""
 
 
@@ -753,7 +761,7 @@ async def test_content_not_matching(httpx_mock: HTTPXMock):
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body2")
+            await client.post("http://test_url", content=b"This is the body2")
         assert (
             str(exception_info.value)
             == """No response can be found for POST request on http://test_url with b'This is the body2' body amongst:
@@ -772,7 +780,7 @@ async def test_headers_and_content_matching(httpx_mock: HTTPXMock):
     )
 
     async with httpx.AsyncClient() as client:
-        response = await client.post("http://test_url", data=b"This is the body")
+        response = await client.post("http://test_url", content=b"This is the body")
         assert response.content == b""
 
 
@@ -788,7 +796,7 @@ async def test_headers_not_matching_and_content_matching(httpx_mock: HTTPXMock):
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -811,7 +819,7 @@ async def test_headers_matching_and_content_not_matching(httpx_mock: HTTPXMock):
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -834,7 +842,7 @@ async def test_headers_and_content_not_matching(httpx_mock: HTTPXMock):
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -854,7 +862,7 @@ async def test_url_and_headers_and_content_matching(httpx_mock: HTTPXMock):
     )
 
     async with httpx.AsyncClient() as client:
-        response = await client.post("http://test_url", data=b"This is the body")
+        response = await client.post("http://test_url", content=b"This is the body")
         assert response.content == b""
 
 
@@ -871,7 +879,7 @@ async def test_headers_not_matching_and_url_and_content_matching(httpx_mock: HTT
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -895,7 +903,7 @@ async def test_url_and_headers_not_matching_and_content_matching(httpx_mock: HTT
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -919,7 +927,7 @@ async def test_url_and_headers_matching_and_content_not_matching(httpx_mock: HTT
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -943,7 +951,7 @@ async def test_headers_matching_and_url_and_content_not_matching(httpx_mock: HTT
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -967,7 +975,7 @@ async def test_url_matching_and_headers_and_content_not_matching(httpx_mock: HTT
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -991,7 +999,7 @@ async def test_url_and_headers_and_content_not_matching(httpx_mock: HTTPXMock):
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1012,7 +1020,7 @@ async def test_method_and_url_and_headers_and_content_matching(httpx_mock: HTTPX
     )
 
     async with httpx.AsyncClient() as client:
-        response = await client.post("http://test_url", data=b"This is the body")
+        response = await client.post("http://test_url", content=b"This is the body")
         assert response.content == b""
 
 
@@ -1032,7 +1040,7 @@ async def test_headers_not_matching_and_method_and_url_and_content_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1059,7 +1067,7 @@ async def test_url_and_headers_not_matching_and_method_and_content_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1086,7 +1094,7 @@ async def test_method_and_url_and_headers_matching_and_content_not_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1113,7 +1121,7 @@ async def test_method_and_headers_matching_and_url_and_content_not_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1140,7 +1148,7 @@ async def test_method_and_url_matching_and_headers_and_content_not_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1167,7 +1175,7 @@ async def test_method_matching_and_url_and_headers_and_content_not_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1194,7 +1202,7 @@ async def test_method_and_url_and_headers_and_content_not_matching(
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            await client.post("http://test_url", data=b"This is the body")
+            await client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1203,3 +1211,47 @@ Match PUT requests on http://test_url2 with {{'user-agent': 'python-httpx/{httpx
 
     # Clean up responses to avoid assertion failure
     httpx_mock.reset(assert_all_responses_were_requested=False)
+
+
+@pytest.mark.asyncio
+async def test_header_as_str_tuple_list(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(
+        headers=[("set-cookie", "key=value"), ("set-cookie", "key2=value2")]
+    )
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value", "key2": "value2"}
+
+
+@pytest.mark.asyncio
+async def test_header_as_bytes_tuple_list(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(
+        headers=[(b"set-cookie", b"key=value"), (b"set-cookie", b"key2=value2")]
+    )
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value", "key2": "value2"}
+
+
+@pytest.mark.asyncio
+async def test_header_as_bytes_dict(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(headers={b"set-cookie": b"key=value"})
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value"}
+
+
+@pytest.mark.asyncio
+async def test_header_as_httpx_headers(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(headers=httpx.Headers({"set-cookie": "key=value"}))
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value"}