+++ tests/test_httpx_sync.py
@@ -432,11 +432,11 @@ def test_requests_retrieval(httpx_mock: HTTPXMock):
     )
 
     with httpx.Client() as client:
-        client.post("http://test_url", data=b"sent content 2")
+        client.post("http://test_url", content=b"sent content 2")
         client.get("http://test_url", headers={"X-TEST": "test header 1"})
-        client.put("http://test_url", data=b"sent content 3")
+        client.put("http://test_url", content=b"sent content 3")
         client.head("http://test_url")
-        client.patch("http://test_url", data=b"sent content 5")
+        client.patch("http://test_url", content=b"sent content 5")
         client.delete("http://test_url", headers={"X-Test": "test header 4"})
 
     assert (
@@ -569,12 +569,15 @@ def test_requests_json_body(httpx_mock: HTTPXMock):
     with httpx.Client() as client:
         response = client.post("http://test_url")
         assert response.json() == {"key 1": "value 1", "key 2": "value 2"}
+        assert response.headers["content-type"] == "application/json"
 
         response = client.get("http://test_url")
         assert response.json() == ["list content 1", "list content 2"]
+        assert response.headers["content-type"] == "application/json"
 
         response = client.put("http://test_url")
         assert response.json() == "string value"
+        assert response.headers["content-type"] == "application/json"
 
 
 def test_callback_raising_exception(httpx_mock: HTTPXMock):
@@ -600,6 +603,7 @@ def test_callback_returning_response(httpx_mock: HTTPXMock):
     with httpx.Client() as client:
         response = client.get("http://test_url")
         assert response.json() == {"url": "http://test_url"}
+        assert response.headers["content-type"] == "application/json"
 
 
 def test_callback_executed_twice(httpx_mock: HTTPXMock):
@@ -611,9 +615,11 @@ def test_callback_executed_twice(httpx_mock: HTTPXMock):
     with httpx.Client() as client:
         response = client.get("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
         response = client.post("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
 
 def test_callback_matching_method(httpx_mock: HTTPXMock):
@@ -625,9 +631,11 @@ def test_callback_matching_method(httpx_mock: HTTPXMock):
     with httpx.Client() as client:
         response = client.get("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
         response = client.get("http://test_url2")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
 
 def test_request_retrieval_with_more_than_one(testdir):
@@ -694,7 +702,7 @@ def test_content_matching(httpx_mock: HTTPXMock):
     httpx_mock.add_response(match_content=b"This is the body")
 
     with httpx.Client() as client:
-        response = client.post("http://test_url", data=b"This is the body")
+        response = client.post("http://test_url", content=b"This is the body")
         assert response.read() == b""
 
 
@@ -703,7 +711,7 @@ def test_content_not_matching(httpx_mock: HTTPXMock):
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body2")
+            client.post("http://test_url", content=b"This is the body2")
         assert (
             str(exception_info.value)
             == """No response can be found for POST request on http://test_url with b'This is the body2' body amongst:
@@ -721,7 +729,7 @@ def test_headers_and_content_matching(httpx_mock: HTTPXMock):
     )
 
     with httpx.Client() as client:
-        response = client.post("http://test_url", data=b"This is the body")
+        response = client.post("http://test_url", content=b"This is the body")
         assert response.content == b""
 
 
@@ -736,7 +744,7 @@ def test_headers_not_matching_and_content_matching(httpx_mock: HTTPXMock):
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -758,7 +766,7 @@ def test_headers_matching_and_content_not_matching(httpx_mock: HTTPXMock):
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -780,7 +788,7 @@ def test_headers_and_content_not_matching(httpx_mock: HTTPXMock):
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -799,7 +807,7 @@ def test_url_and_headers_and_content_matching(httpx_mock: HTTPXMock):
     )
 
     with httpx.Client() as client:
-        response = client.post("http://test_url", data=b"This is the body")
+        response = client.post("http://test_url", content=b"This is the body")
         assert response.content == b""
 
 
@@ -815,7 +823,7 @@ def test_headers_not_matching_and_url_and_content_matching(httpx_mock: HTTPXMock
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -838,7 +846,7 @@ def test_url_and_headers_not_matching_and_content_matching(httpx_mock: HTTPXMock
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -861,7 +869,7 @@ def test_url_and_headers_matching_and_content_not_matching(httpx_mock: HTTPXMock
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -884,7 +892,7 @@ def test_headers_matching_and_url_and_content_not_matching(httpx_mock: HTTPXMock
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -907,7 +915,7 @@ def test_url_matching_and_headers_and_content_not_matching(httpx_mock: HTTPXMock
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -930,7 +938,7 @@ def test_url_and_headers_and_content_not_matching(httpx_mock: HTTPXMock):
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -950,7 +958,7 @@ def test_method_and_url_and_headers_and_content_matching(httpx_mock: HTTPXMock):
     )
 
     with httpx.Client() as client:
-        response = client.post("http://test_url", data=b"This is the body")
+        response = client.post("http://test_url", content=b"This is the body")
         assert response.content == b""
 
 
@@ -969,7 +977,7 @@ def test_headers_not_matching_and_method_and_url_and_content_matching(
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -995,7 +1003,7 @@ def test_url_and_headers_not_matching_and_method_and_content_matching(
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1021,7 +1029,7 @@ def test_method_and_url_and_headers_matching_and_content_not_matching(
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1047,7 +1055,7 @@ def test_method_and_headers_matching_and_url_and_content_not_matching(
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1073,7 +1081,7 @@ def test_method_and_url_matching_and_headers_and_content_not_matching(
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1099,7 +1107,7 @@ def test_method_matching_and_url_and_headers_and_content_not_matching(
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1123,7 +1131,7 @@ def test_method_and_url_and_headers_and_content_not_matching(httpx_mock: HTTPXMo
 
     with httpx.Client() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
-            client.post("http://test_url", data=b"This is the body")
+            client.post("http://test_url", content=b"This is the body")
         assert (
             str(exception_info.value)
             == f"""No response can be found for POST request on http://test_url with {{'host': 'test_url', 'user-agent': 'python-httpx/{httpx.__version__}'}} headers and b'This is the body' body amongst:
@@ -1132,3 +1140,43 @@ Match PUT requests on http://test_url2 with {{'user-agent': 'python-httpx/{httpx
 
     # Clean up responses to avoid assertion failure
     httpx_mock.reset(assert_all_responses_were_requested=False)
+
+
+def test_header_as_str_tuple_list(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(
+        headers=[("set-cookie", "key=value"), ("set-cookie", "key2=value2")]
+    )
+
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value", "key2": "value2"}
+
+
+def test_header_as_bytes_tuple_list(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(
+        headers=[(b"set-cookie", b"key=value"), (b"set-cookie", b"key2=value2")]
+    )
+
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value", "key2": "value2"}
+
+
+def test_header_as_bytes_dict(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(headers={b"set-cookie": b"key=value"})
+
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value"}
+
+
+def test_header_as_httpx_headers(httpx_mock: HTTPXMock):
+    httpx_mock.add_response(headers=httpx.Headers({"set-cookie": "key=value"}))
+
+    with httpx.Client() as client:
+        response = client.get("http://test_url")
+
+    assert dict(response.cookies) == {"key": "value"}